+++ src/main/java/org/informatics/service/ShopService.java
@@ -5,15 +5,14 @@ import org.informatics.data.Goods;
 import org.informatics.data.Receipt;
 import org.informatics.data.Shop;
 
-import java.util.Map;
-import java.util.Set;
+import java.math.BigDecimal;
 
 public interface ShopService {
     public void addCashiers(Shop shop, Cashier cashier);
     public void addDeliveredProduct(Shop shop,Goods goods);
     public void addSoldProduct(Shop shop,Goods goods);
     public void addReceipt(Shop shop,Receipt receipt);
-    public double calculateExpenses(Shop shop);
-    public double calculateIncome(Shop shop);
-    public double calculateProfit(Shop shop);
+    public BigDecimal calculateExpenses(Shop shop);
+    public BigDecimal calculateIncome(Shop shop);
+    public BigDecimal calculateProfit(Shop shop);
 }
